pkg/karmadactl: unit test join
In this commit, we unit test the process of joining a cluster
to the Karmada control plane. The tests validate the arguments
passed to the join cluster operation and ensure the creation of
the new cluster runtime object. Additionally, we verify that the
necessary access secrets for the new cluster in push mode are correctly
generated in the Karmada control plane.

Signed-off-by: Mohamed Awnallah <mohamedmohey2352@gmail.com>